Move all the cache tests in a single file for readability
